refactor : GameResult 패키지 user -> result로 변경
