+++ src/main/java/domain/result/GameResult.java
@@ -1,4 +1,4 @@
-package domain.user;
+package domain.result;
 
 import java.util.Arrays;
