+++ src/main/java/view/OutputView.java
@@ -69,4 +69,6 @@ public class OutputView {
         System.out.println("딜러는 16이하라 한장의 카드를 더 받았습니다.");
     }
 
+    public static void printProfits(Players players, Dealer dealer) {
+    }
 }
