+++ src/test/java/domain/user/GameResultTest.java
@@ -1,5 +1,6 @@
 package domain.user;
 
+import domain.result.GameResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
