+++ src/test/java/domain/user/PlayerTest.java
@@ -3,6 +3,7 @@ package domain.user;
 import domain.card.Card;
 import domain.card.Symbol;
 import domain.card.Type;
+import domain.result.GameResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
